Email on payments

+++ client_rest_api/apps/payment/views.py
@@ -1813,29 +1813,94 @@ class CancelWithdrawalRequest(APIView):
                 response['httpstatus'] = status.HTTP_400_BAD_REQUEST
                 return Response(response, status=response.get('httpstatus'))
             
-            withdrawalRes = CRM()
-            res = withdrawalRes.cancel_withdrawal(withdrawalID=transId)
-
-
-            if not res['success']:
-                response['status'] = 'error'
-                response['errorcode'] = status.HTTP_400_BAD_REQUEST
-                response['reason'] = "Unable to cancel pending request!!!"
-                response['httpstatus'] = status.HTTP_400_BAD_REQUEST
-                return Response(response, status=response.get('httpstatus'))
-            
-
             order = OrderDetails.objects.filter(brokerBankingId=transId).first()
 
             if order:
-                WithdrawalApprovals.objects.filter(ordertransactionid=order).delete()
-                order.delete()
-                response['result'] = "Withdrawal Request Cancelled Successfully!!!"
-                return Response(response, status=response.get('httpstatus'))
+                withRes = WithdrawalApprovals.objects.filter(ordertransactionid=order).first()
+                if not withRes:
+                    response['status'] = 'error'
+                    response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                    response['reason'] = "Withdrawal Request Not Found!!!"
+                    response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                    return Response(response, status=response.get('httpstatus'))
+
+                if withRes.first_approval_action:
+                    response['status'] = "error"
+                    response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                    response['result'] = "Withdrawal Request is under process!!!"
+                    response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                    return Response(response, response.get('httpstatus'))
+                
+                else:
+                    withdrawalRes = CRM()
+                    res = withdrawalRes.cancel_withdrawal(withdrawalID=transId)
+                    if not res['success']:
+                        response['status'] = 'error'
+                        response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                        response['reason'] = "Unable to cancel pending request!!!"
+                        response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                        return Response(response, status=response.get('httpstatus'))
+                    else:
+                        WithdrawalApprovals.objects.filter(ordertransactionid=order).delete()
+                        order.delete()
+                        response['result'] = "Withdrawal Request Cancelled Successfully!!!"
+                        return Response(response, status=response.get('httpstatus'))
             else:
                 response['reason'] = "Withdrawal Request"
                 response['result'] = "Withdrawal Request Cancelled Successfully!!!"
                 return Response(response, status=response.get('httpstatus'))
+
+        except Exception as e:
+            print(f"Error in cancelling the Withdrawal Request: {str(e)}")
+            response['status'] = 'error'
+            response['errorcode'] = status.HTTP_400_BAD_REQUEST
+            response['reason'] = str(e)
+            response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+            return Response(response, status=response.get('httpstatus'))
+
+
+class HideWithdarwalRequest(APIView):
+
+    def get(self, request):
+        try:
+            response = {"status":"success", "errorcode": "", "reason": "", "result": "", "httpstatus": status.HTTP_200_OK}
+
+            transId = request.query_params.get('transId')
+            
+            if not transId:
+                response['status'] = 'error'
+                response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                response['reason'] = "Transaction Id is required!!!"
+                response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                return Response(response, status=response.get('httpstatus'))
+            
+            order = OrderDetails.objects.filter(brokerBankingId=transId).first()
+
+            if order:
+                withRes = WithdrawalApprovals.objects.filter(ordertransactionid=order).first()
+
+                if not withRes:
+                    response['status'] = 'error'
+                    response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                    response['reason'] = "Withdrawal Request Not Found!!!"
+                    response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                    return Response(response, status=response.get('httpstatus'))
+
+                if withRes.first_approval_action:
+                    response['status'] = "error"
+                    response['errorcode'] = status.HTTP_400_BAD_REQUEST
+                    response['result'] = "Withdrawal Request is under process!!!"
+                    response['httpstatus'] = status.HTTP_400_BAD_REQUEST
+                    return Response(response, response.get('httpstatus'))
+                
+                else:
+                    response['result'] = "Withdrawal request pending!!!"
+                    return Response(response, response.get('httpstatus'))
+                
+            else:
+                response['result'] = "No Order Found!!!"
+                return Response(response, status=response.get('httpstatus'))
+
 
         except Exception as e:
             print(f"Error in cancelling the Withdrawal Request: {str(e)}")
